完成下单业务逻辑

+++ shop-pojo/src/main/java/com/ling/shop/pojo/TradeOrder.java
@@ -54,7 +54,7 @@ public class TradeOrder implements Serializable {
     /**
      * 商品总价
      */
-    private Long goodsAmount;
+    private BigDecimal goodsAmount;
     /**
      * 运费
      */
